+++ src/tests/ExpressionTest.java
@@ -18,7 +18,7 @@ public class ExpressionTest {
 		
 		exception.expect(Exception.class);
 		exception.expectMessage("Invalid objective.");
-		e = new Expression(Expression.OBJECTIVE_MAX+1, new int[0], new int[0][0], new int[0], new int [0]);	
+		e = new Expression(Expression.OBJECTIVE_MAX+1, new double[0], new double[0][0], new double[0], new double [0]);	
 	}
 	
 	@Test
@@ -27,7 +27,7 @@ public class ExpressionTest {
 		
 		exception.expect(Exception.class);
 		exception.expectMessage("Invalid objective.");
-		e = new Expression(Expression.OBJECTIVE_MIN-1, new int[0], new int[0][0], new int[0], new int [0]);	
+		e = new Expression(Expression.OBJECTIVE_MIN-1, new double[0], new double[0][0], new double[0], new double [0]);	
 	}
 	
 	@Test
@@ -36,7 +36,7 @@ public class ExpressionTest {
 		
 		exception.expect(Exception.class);
 		exception.expectMessage("Neither objective function, constratints nor constratint signs length should be 0.");
-		e = new Expression(Expression.OBJECTIVE_MAX, new int[0], new int[1][1], new int[1], new int [1]);	
+		e = new Expression(Expression.OBJECTIVE_MAX, new double[0], new double[1][1], new double[1], new double [1]);	
 	}
 	
 	@Test
@@ -45,7 +45,7 @@ public class ExpressionTest {
 		
 		exception.expect(Exception.class);
 		exception.expectMessage("Neither objective function, constratints nor constratint signs length should be 0.");
-		e = new Expression(Expression.OBJECTIVE_MAX, new int[1], new int[0][0], new int[1], new int [1]);	
+		e = new Expression(Expression.OBJECTIVE_MAX, new double[1], new double[0][0], new double[1], new double [1]);	
 	}
 	
 	@Test
@@ -54,7 +54,7 @@ public class ExpressionTest {
 		
 		exception.expect(Exception.class);
 		exception.expectMessage("Neither objective function, constratints nor constratint signs length should be 0.");
-		e = new Expression(Expression.OBJECTIVE_MAX, new int[1], new int[1][0], new int[0], new int [1]);	
+		e = new Expression(Expression.OBJECTIVE_MAX, new double[1], new double[1][0], new double[0], new double [1]);	
 	}
 	
 	@Test
@@ -63,7 +63,7 @@ public class ExpressionTest {
 		
 		exception.expect(Exception.class);
 		exception.expectMessage("Neither objective function, constratints nor constratint signs length should be 0.");
-		e = new Expression(Expression.OBJECTIVE_MAX, new int[1], new int[1][0], new int[1], new int [0]);	
+		e = new Expression(Expression.OBJECTIVE_MAX, new double[1], new double[1][0], new double[1], new double [0]);	
 	}
 	
 	@Test
@@ -72,7 +72,7 @@ public class ExpressionTest {
 		
 		exception.expect(Exception.class);
 		exception.expectMessage("Number of decision variables or constraints incompatible.");
-		e = new Expression(Expression.OBJECTIVE_MAX, new int[2], new int[1][0], new int[1], new int [1]);
+		e = new Expression(Expression.OBJECTIVE_MAX, new double[2], new double[1][0], new double[1], new double [1]);
 	}
 	
 	@Test
@@ -81,7 +81,7 @@ public class ExpressionTest {
 		
 		exception.expect(Exception.class);
 		exception.expectMessage("Number of decision variables or constraints incompatible.");
-		e = new Expression(Expression.OBJECTIVE_MAX, new int[2], new int[2][1], new int[2], new int [2]);
+		e = new Expression(Expression.OBJECTIVE_MAX, new double[2], new double[2][1], new double[2], new double [2]);
 	}
 	
 	@Test
@@ -90,7 +90,7 @@ public class ExpressionTest {
 		
 		exception.expect(Exception.class);
 		exception.expectMessage("Number of decision variables or constraints incompatible.");
-		e = new Expression(Expression.OBJECTIVE_MAX, new int[2], new int[3][2], new int[2], new int [3]);
+		e = new Expression(Expression.OBJECTIVE_MAX, new double[2], new double[3][2], new double[2], new double [3]);
 	}
 	
 	@Test
@@ -99,7 +99,7 @@ public class ExpressionTest {
 		
 		exception.expect(Exception.class);
 		exception.expectMessage("Number of decision variables or constraints incompatible.");
-		e = new Expression(Expression.OBJECTIVE_MAX, new int[2], new int[3][2], new int[3], new int [2]);
+		e = new Expression(Expression.OBJECTIVE_MAX, new double[2], new double[3][2], new double[3], new double[2]);
 	}
 	
 	@Test
@@ -108,16 +108,16 @@ public class ExpressionTest {
 		
 		exception.expect(Exception.class);
 		exception.expectMessage("Invalid constraint sign.");
-		int[] constraintSign = {Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_LT+1, Expression.CONSTRAINT_SIGN_GT};
-		e = new Expression(Expression.OBJECTIVE_MAX, new int[2], new int[3][2], constraintSign, new int [3]);
+		double[] constraintSign = {Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_LT+1, Expression.CONSTRAINT_SIGN_GT};
+		e = new Expression(Expression.OBJECTIVE_MAX, new double[2], new double[3][2], constraintSign, new double[3]);
 	}
 
 	private Expression getDefaultExpression() throws Exception {
 		int objective = Expression.OBJECTIVE_MAX;
-		int[] objectiveFunction = {80, 60};
-		int[][] constraints = {{4, 6}, {4, 2}, {0, 1}};
-		int[] constraintSigns = {Expression.CONSTRAINT_SIGN_GT, Expression.CONSTRAINT_SIGN_LT, Expression.CONSTRAINT_SIGN_LT};
-		int[] b = {24, 16, 3};
+		double[] objectiveFunction = {80, 60};
+		double[][] constraints = {{4, 6}, {4, 2}, {0, 1}};
+		double[] constraintSigns = {Expression.CONSTRAINT_SIGN_GT, Expression.CONSTRAINT_SIGN_LT, Expression.CONSTRAINT_SIGN_LT};
+		double[] b = {24, 16, 3};
 
 		return new Expression(objective, objectiveFunction, constraints, constraintSigns, b);
 	}
